chore: fixed merge conflicts

+++ app/src/main/java/com/mafaly/moviematch/services/MovieService.kt
@@ -2,7 +2,6 @@ package com.mafaly.moviematch.services
 
 import android.content.Context
 import com.mafaly.moviematch.db.AppDatabase
-import com.mafaly.moviematch.db.entities.GameEntity
 import com.mafaly.moviematch.db.entities.MovieEntity
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
@@ -11,7 +10,7 @@ import kotlinx.coroutines.withContext
 class MovieService {
 
     companion object {
-        suspend fun getMovieById(context: Context, movieId: Int): MovieEntity? {
+        suspend fun getMovieById(context: Context, movieId: Long): MovieEntity? {
             return withContext(Dispatchers.IO) {
                 val appDatabase = AppDatabase.getInstance(context)
                 val deferredDuel = async { appDatabase.movieDao().getMovieById(movieId) }
